requirements.txt
beautifulsoup4==4.10.0
pandas==1.3.4
python-dotenv==0.19.2
python_dateutil==2.8.2
selenium==4.1.0
